+++ app/src/lib/build-info.js
@@ -1,2 +1,2 @@
-export const APP_BUILD = 'v1.3.1'
-export const BUILD_DATE = '2026-06-05 13:10'
+export const APP_BUILD = 'v1.3.3'
+export const BUILD_DATE = '2026-06-05 15:30'